third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_lostpointercapture_remove_setcapture_node-expected.txt

This is a testharness.js-based test.
[FAIL] setPointerCapture target removed by lostpointercapture
  assert_equals: expected "pointerover@button, pointerenter@button, pointermove@button, pointerdown@button, pointerout@button, pointerleave@button, pointerover@target0, pointerenter@target0, gotpointercapture@target0, pointermove@target0, lostpointercapture@target0, pointerout@target0, pointerleave@target0, pointerover@button, pointerenter@button, pointermove@button, pointerup@button" but got "pointerover@button, pointerenter@button, pointermove@button, pointerdown@button, pointerout@button, pointerleave@button, pointerover@target0, pointerenter@target0, gotpointercapture@target0, pointermove@target0, lostpointercapture@target0, pointerout@target0, pointerleave@target0, pointerup@target1"
Harness: the test ran to completion.
